+++ godot-src/export_presets.cfg
@@ -26,7 +26,11 @@ vram_texture_compression/for_desktop=true
 vram_texture_compression/for_mobile=false
 html/export_icon=true
 html/custom_html_shell=""
-html/head_include="<script src=\"microsoft.cognitiveservices.speech.sdk.bundle.js\"></script><script src=\"script.js\"></script>"
+html/head_include=
+"
+<script src=\"microsoft.cognitiveservices.speech.sdk.bundle.js\"></script>
+<script type=\"module\" src=\"script.js\"></script>
+"
 html/canvas_resize_policy=2
 html/focus_canvas_on_start=true
 html/experimental_virtual_keyboard=false
