--- index.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>TLBTC</title>
-</head>
-<body>
-    <center> 
-        <iframe src="godot_game.html" frameborder="0" width="800" height="600" ></iframe>
-    </center>
-</body>
-</html>
