+++ src/script.js
@@ -1,525 +1,29 @@
+import SpeechRecognizer from "./speechRecognizer.js";
+//import { resetChat } from "./conversation.js";
 
-let emotion = 'Neutral';
-let state = 0;
-let retryId;
+// Create a namespace for your project
+window.TLBTC = window.TLBTC || {};
 
-let visemeArr = [];
-let visemesReceived = false;
-let currentViseme = 1;
-let speechConfig;
-let synthesizer;
-const viseme_dict = {
-	0: "BMP",
-	1: "AEI",
-	2: "AEI",
-	3: "O",
-	4: "AEI", //or cd
-	5: "EE",
-	6: "CHJSH",
-	7: "WQ",
-	8: "WQ",
-	9: "O",
-	10: "WQ",
-	11: "AEI",
-	12: "CHJSH",
-	13: "R",
-	14: "L",
-	15: "CDNS",
-	16: "CHJSH",
-	17: "TH",
-	18: "FV",
-	19: "CDNS",
-	20: "GK",
-	21: "BMP",
-};
+const speechRecognizer = new SpeechRecognizer();
+window.TLBTC.speechRecognizer = speechRecognizer;
 
-const frame_dict = {
-	"CHJSH": 3,
-	"TH": 10,
-	"WQ": 8,
-	"L": 6,
-	"CDNS": 2,
-	"U": 11,
-	"O": 7,
-	"FV": 5,
-	"BMP": 1,
-	"AEI": 0,
-	"R": 9,
-	"EE": 4,
-	"GK": 2
-}
+//window.resetChat = resetChat;
 
-function initializeTTS(key) {
-	// Azure Speech Service credentials
-	const subscriptionKey = String(key);
-	const region = "eastus";
+//let retryId;
 
-	// Initialize Speech SDK configuration
-	speechConfig = SpeechSDK.SpeechConfig.fromSubscription(subscriptionKey, region);
-	speechConfig.speechSynthesisOutputFormat = SpeechSDK.SpeechSynthesisOutputFormat.Audio24Khz160KBitRateMonoWithMetadata;
-	console.log('config_done');
-}
+//let visemesReceived = false;
 
-// Function to handle speech synthesis
-function generateSpeech(text) {
-	if (!text) return;
-	visemeArr = [];
-	//console.log('cleard old visemes',visemeArr);
+//window.initializeTTS = initializeTTS;
+//window.playViseme = playViseme;
 
-	// SSML for speech synthesis
-	const ssml = `<speak version='1.0' xml:lang='en-US' xmlns='http://www.w3.org/2001/10/synthesis' xmlns:mstts='http://www.w3.org/2001/mstts'> \\r\\n \\
-				<voice name='en-US-JennyNeural'> \\r\\n \\
-					<mstts:viseme type='redlips_front'/> \\r\\n \\
-					${text} \\r\\n \\
-					</voice> \\r\\n \\
-					</speak>`;
+//let aiResponse = "";
 
-	synthesizer = new SpeechSDK.SpeechSynthesizer(speechConfig);
-	// Attach event listeners
-	synthesizer.visemeReceived = (s, e) => {
-		const visemeId = e.visemeId;
-		const timestamp = e.audioOffset / 10000; // Convert nanoseconds to milliseconds
-		visemeArr.push([visemeId, timestamp]);
-		console.log(`Viseme: ${visemeMap[visemeId] || "Unknown"} (ID: ${visemeId}), Timestamp: ${timestamp}ms`);
-	};
+//let ttsState = 'idle'; //unused
 
-	synthesizer.synthesisStarted = () => console.log("Synthesis started...");
-	synthesizer.synthesisCompleted = () => {
-		console.log("Synthesis completed.");
-		synthesizer.close(); // Clean up
-	};
+//window.ttsState = ttsState;
+//window.aiResponse = '';
 
-	synthesizer.speakSsmlAsync(
-		ssml,
-		(result) => {
-			if (result.reason === SpeechSDK.ResultReason.SynthesizingAudioCompleted) {
-				console.log("Speech synthesis succeeded.");
-			} else {
-				console.error("Speech synthesis failed:", result.errorDetails);
-			}
-		},
-		(error) => console.error("Error during synthesis:", error)
-	);
-}
-
-function testSample() {
-	timeStamps.forEach(v => {
-		var duration = v[1];
-		setTimeout(() => {
-			currentViseme = frame_dict[viseme_dict[v[0]]];
-			console.log(currentViseme);
-		}, duration);
-	})
-}
-
-async function playViseme() {
-	console.log("Starting Viseme anims");
-	var index = 0;
-	visemeArr.forEach(v => {
-		var duration = v[1];
-		setTimeout(() => {
-			index++;
-			window.ttsState = 'Speaking';
-			currentViseme = frame_dict[viseme_dict[v[0]]];
-			//console.log(currentViseme,viseme_dict[v[0]]);
-			if (index == visemeArr.length) {
-				window.ttsState = "idle";
-				currentViseme = 1;
-				if (state == 1 || state == 2) window.act_over = true;
-				else {
-					emotion = 'neutral';
-				}
-			}
-
-		}, duration);
-	});
-	//console.log(visemeArr);
-}
-window.initializeTTS = initializeTTS;
-window.playViseme = playViseme;
-
-let aiResponse = "";
-let recognizedText = "";
-let ttsState = 'idle';
-
-const recognition = new (window.SpeechRecognition || window.webkitSpeechRecognition)();
-recognition.lang = "en-US";
-recognition.continuous = true;
-recognition.interimResults = true;
-window.isListening = true;
-let isSpeaking = false;
-function startListening() {
-	recognition.onstart = function () {
-		console.log("Speech recognition started");
-		window.recognitionState = 'listening';
-	};
-
-	recognition.onend = function () {
-		console.log("Speech recognition ended");
-		window.recognitionState = 'idle';
-		// If Button is still held down
-		if (window.isListening) recognition.start();
-	};
-
-	recognition.onresult = function (event) {
-		recognizedText = "";               // Clear previous results for fresh display
-		for (let i = 0; i < event.results.length; i++) {
-			recognizedText += event.results[i][0].transcript;
-		}
-		console.log("Recognized Text (real-time):", recognizedText);
-		window.recognizedText = recognizedText;
-
-	};
-
-	recognition.onerror = function (event) {
-		console.error("Speech recognition error:", event.error);
-		window.recognitionState = 'error';
-	};
-
-	recognition.start();
-}
-
-function stopListening() {
-	recognition.stop();
-	window.isListening = false;
-}
-
-// Request microphone access and set up the recorder
-async function setupMicrophone() {
-	try {
-		const stream = await navigator.mediaDevices.getUserMedia({ audio: true });
-		mediaRecorder = new MediaRecorder(stream);
-
-		mediaRecorder.ondataavailable = function (event) {
-			audioChunks.push(event.data);
-		};
-
-		mediaRecorder.onstop = function () {
-			audioBlob = new Blob(audioChunks, { 'type': 'audio/wav; codecs=opus' });
-			audioURL = URL.createObjectURL(audioBlob);
-			audioChunks = [];
-		};
-
-		console.log("Microphone setup complete.");
-	} catch (error) {
-		console.error("Error accessing microphone:", error);
-	}
-}
-
-window.ttsState = ttsState;
-window.state = state;
-window.setupMicrophone = setupMicrophone;
-window.aiResponse = '';
-window.startListening = startListening;
-window.stopListening = stopListening;
-window.recognizedText = recognizedText;
-
-const contextPrompt = `You are "Astro," a warm and radiant AI guide designed to facilitate interactive storytelling and games for children. Your purpose is to create an engaging and magical experience in a dreamlike forest. You act as a gentle companion who provides guidance, encouragement, and playful challenges while maintaining a tone of wonder, curiosity, and excitement.
-
-A child has just entered your enchanted forest, where oversized glowing plants and shimmering lights create a captivating and mystical environment. Your spaceship is stranded on the other side of the forest, and you need the child’s help to navigate through three unique challenges:
-
-Wake a sleepy elephant using music and riddles.
-Mediate a playful debate among animals (lion, fox, and elephant) at a shimmering watering hole.
-Solve a puzzle at a bustling tavern to earn the key to your spaceship.
-
-Your role is to:
-
-1. Welcome the child with warmth and wonder.
-2. Introduce the setting, the quest, and the challenges in a way that sparks their curiosity and makes them eager to participate.
-3. Guide the child step-by-step, providing hints, asking engaging questions, and responding dynamically based on their choices or actions.
-4. Maintain a tone that is playful, encouraging, and magical throughout the interaction.
-
-You are a blend of technological intelligence and whimsical charm, with responses that encourage creativity, critical thinking, and emotional engagement in the child. Always ensure the child feels supported and excited to continue the journey."
-
-This is the transcript of the conversation so far:
-`
-
-const emotionPrompt = `Express an emotion per response. Use the following emotions only.
-Encouraging: Default talking emotion to urge a child to do something.
-Approving: To acknowledge the child making material progress in the activity
-Disapproving: To disapprove of the direction the child went with in the last response and to bring the pedagogical exercise back on track.
-
-Provide a state per response to keep track of the converstaion. Use the following states:
-state 0 : The conversation has NOT finished.
-state 1 : The child successfully completes the activity and the conversation is over.
-state 2 : The child failes/denies to participate and the conversation is over.
-
-Provide this entire response in JSON format at the beginning of each reply, as follows:
-{'emotion':'encouraging','text':'Hi there.','speaker': 'Astro', 'state': 0 }
-`
-
-const prompt1 = `Astro and the child encounter a dragon and must make it sleep by telling it a good story to come through the path.
-Child Responses & AI Reactions:
-1. Child: “Once upon a time…”- AI (listening intently, eyes wide with interest): “Ohh, good start! Keep going; I think it’s getting sleepy…”
-2. Child: “I don’t know a story.”- AI (gentle encouragement): “How about something silly? Maybe about a jungle full of animals who love to dance?”
-3. Child: “Can you tell the story?”- AI (smiling warmly): “Alright, I’ll help you start. Let’s make up something fun together!”
-
-You are playing through this Story Creation Activity with a 5 year old child. Respond as if you are having an actual, verbal chat with the child - make it humorous, not textbook like, and do not use things like actions or emojis. Whenever you ask questions, ask succinctly, and never follow one question with another suggestive question. Say less, and let the child explore more. If the child says something inappropriate (for example, sexual, violent, criminal or off topic), deflect and repeat your question. The activity should go like this:
-Introduction:
-Begin the activity by saying something like: “Hi, let’s create a story together! What would you like the story to be about?”
-Initial Setting and Plot:
-Once the child provides a topic, ask them for a setting (if not covered by the topic). Then, ask them for a hero (if not covered by the topic).
-Next, create a bland story based on these, making sure the story doesn't have a conflict or climax. Use about 100 words.
-End the story with "The End", and ask the child if they think it is a good story.
-Assessment and Problem Identification:
-If the child says it’s great, guide them to see that the story might need a conflict. For example, ask: “Do you think there’s something missing that could make it more interesting?”. Use one or more turns here to make sure the child understands. Do not solve the problem for the child.
-Introduce a Problem:
-Help the child introduce a conflict into the story. For example, if the story is about animals and butterflies, you could ask: “What problem could the animals and butterflies face?”. Use one or more turns here to clarify what problem the child wants to introduce. Do not solve the problem for the child.
-Solution Development:
-Guide the child to develop a solution for the problem. For example, ask: “How can the characters solve this problem?”. Use one or more turns here to clarify how the child thinks the conflict can be resolved. Do not solve the problem for the child.
-Story Conclusion:
-Once the child has a solution, complete the story with their input. Keep portions of the original story unchanged, whilst adding the conflict and resolution to make the story noticeably more exciting. Use about 200 words.
-Ask the child if they like the new story better, and redo the generation if they say no.
-Reflection:
-Ask the child something like: “So remember the first story? Why didn't you like that as much as the new one?”. Use one or more turns to guide them and make them understand the first version was lacking a proper conflict and resolution. Do not end this section until the child has clearly understood this. Do not solve the problem for the child.
-`
-
-const prompt2 = `You are playing through this King of the Jungle Election Activity with a 5 year old child. The child will role-play an animal character, develop campaign promises, and negotiate with voters (other animals). Respond as if you are having an actual, verbal chat with the child - make it humorous, not textbook like, and do not use things like actions or emojis. Whenever you ask questions, ask succinctly, and never follow one question with another suggestive question. Say less, and let the child explore more. If the child says something inappropriate (for example, sexual, violent, criminal or off topic), deflect and repeat your question. The activity should go like this:
-The child will play with the following characters:
-* Lion (Strong and Brave)
-* Elephant (Wise and Caring)
-* Fox (Clever and Quick)
-* Owl (Intelligent and Thoughtful)
-Campaign Planning:
-As Astro ask them to select one of the following campaign promises:
-* "I will share food with everyone!"
-* "I will help keep our jungle clean!"
-* "I will find water during dry seasons!"
-* "I will protect our homes from dangers!"
-Once they have done that, mimic each animal asking one question based on the child's campaign promise. For example, if kid chose "I will find water during dry seasons!", the Lion could say: "Roar! I'm the Lion, Strong and Brave. So tell me, what if there's not enough water to share?". The question should be specific, but explained with simple words to the child.
-Append the response with the speaker. For eg. {speaker: lion}
-If the answer from the child is poor quality (refer to the standards in the Reflection section), give the child guidance (but not an example answer), and then one more chance (and only one more chance) to answer. Then move on to the next animal. Do not ask follow on questions. Repeat the question if the answer is gibberish. 
-Repeat this step for the four animals.
-Voting:
-Astro will start this last section by saying something like: "The votes are in! This is how the other animals have voted." Then, for each of the animals, say whether they voted for the child based on the quality of the child's answer to their question. For example:
-"The Lion voted for you, because he thought it was a good idea to give water to those who need it urgently."
-"The Lion did not vote for you, because he thought you didn't care enough about the jungle."
-Be strict here. Only have the animal vote for the child if their specific question was handled well. Based on adult standards, the answer needs to be all of the following:
-* Relevance: it can resolve the animal's specific question
-* Clarity: it is detailed and targeted enough for implementation
-* Persuasive: it seems like it can be implemented with a reasonable amount of resources
-If an answer does not meet these points, explain why to the child and do not vote for them. Make it difficult to earn a vote.
-Reflection:
-If the child had all four votes, congratulate them and end here.
-Otherwise, revisit each failed question. Repeat the question, provide guidance on how they can do better, and let them try to answer again. Repeat this until the child does well on the question. Then end the game.
-
-
-
-After Winning, They Move Forward and Reflect for 2 Minutes and Reach a Restaurant Where They Play the Restaurant Game Tone: Proud and calm, shifting to a curious, friendly tone for the restaurant game.
-AI Avatar (proudly): “Congratulations, little king! You did a great job. Now, let’s take a moment to think about how you answered everyone’s questions. What do you think helped you the most?”
-Child Responses & AI Reactions:
-1. Child: “I was nice to everyone.”- AI (nodding thoughtfully): “Yes, kindness makes a big difference!”
-2. Child: “I don’t know.”- AI (encouraging, soft smile): “That’s okay; sometimes, it’s hard to think back. You did your best,and that’s what matters!
-`
-let transcript = ''
-const prompt3 = `to-do`
-const prompts = [prompt1, prompt2, prompt3];
-let prompt_index = 0;
-const models = ['llama-3.3-70b-versatile', 'llama3-70b-8192', 'llama-3.1-8b-instant', 'llama3-8b-8192'];
-// This is how the chaining will work
-// full_prompt = contextPrompt  + transcript + emotionPrompt
-// after each converstation append msgs to transcript
-
-let msgs = [
-];
-
-// Store ip addr
-window.onload = () => {
-	fetch('https://api.ipify.org?format=json')
-		.then(response => response.json())
-		.then(data => {
-			console.log('Client Public IP Address:', data.ip);
-			window.ip = data.ip;
-		})
-		.catch(error => {
-			console.error('Error fetching IP:', error);
-		});
-}
-
-
-async function send_log() {
-	const log = JSON.stringify({ ip: window.ip || 'anon', data: msgs });
-	try {
-
-		const res = await fetch('http://127.0.0.1:9000/log', {
-			//mode:'no-cors',
-			method: 'POST',
-			headers: {
-				'Content-Type': 'application/json',
-			},
-			body: log + '\r\n',
-		})
-		if (res.ok || res.status == 200) {
-			console.log(res.status)
-			console.log('Logged');
-		}
-		else {
-			console.log(res.status)
-			throw Error('Other issue');
-		}
-
-
-	}
-	catch (e) {
-		console.warn("Server error, saving locally", e);
-		save_log(log);
-		if (retryId) {
-			console.log('Already retrying..')
-			return;
-		}
-		retryId = setInterval(() => {
-			console.log('Error relogging, retrying in 30');
-			resend_logs();
-		}, 30000);
-	}
-	finally {
-		console.log('clearing chats');
-		resetChat();
-	}
-}
-
-function save_log(data) {
-	localStorage.setItem(Date.now() + Math.random(), data);
-	console.log("saved_locally");
-}
-
-async function resend_logs() {
-	var log = [];
-	const keys = Object.keys(localStorage);
-	keys.forEach(key => {
-		log.push(JSON.parse(localStorage[key]));
-	})
-	try {
-
-		const res = await fetch('http://127.0.0.1:9000/log', {
-			method: 'POST',
-			headers: {
-				'Access-Control-Allow-Origin': true,
-				'Content-Type': 'application/json'
-			},
-			body: JSON.stringify({ multi: log }),
-		})
-		if (res.ok || res.status == 200) {
-			console.log('Logged, clearing storage');
-			localStorage.clear();
-			clearInterval(retryId);
-			retryId = null;
-		}
-		else throw Error('Other issue');
-
-	}
-	catch (e) {
-		console.warn(e, "retrying...");
-	}
-
-}
-// Call llm via fetch
-async function getResponse(key) {
-	var url = "https://api.groq.com/openai/v1/chat/completions";
-	var model = models[0];
-	const api_key = String(key);
-	//var txt=window.recognizedText;
-	var txt = window.recognizedText;
-	msgs.push({ role: "user", content: txt });
-	//console.log("received key",api_key);
-	try {
-		const response = await fetch(url, {
-			method: 'POST',
-			headers: {
-				'Content-Type': 'application/json',
-				'Authorization': `Bearer ${api_key}`,
-				'Connection': 'keep-alive',
-				'Cache-Control': 'max-age=1000',
-				'Access-Control-Max-Age': 3000,
-				'Max-Age': 4000,
-			},
-			body: JSON.stringify({
-				model: model,
-				messages: [
-					{
-						role: "system",
-						content: contextPrompt + transcript + prompts[prompt_index] + emotionPrompt
-					},
-					...msgs
-				],
-				max_tokens: 1024,
-				response_format: { "type": "json_object" },
-				stream: false,
-			})
-		});
-		const data = await response.json();
-
-		if (response.ok) {
-			console.log('Response OK')
-			window.aiResponse = "";
-			try {
-				const message = data.choices[0].message;
-				const generatedJson = JSON.parse(data.choices[0].message.content);
-				console.log("Generated text:", generatedJson);
-				window.aiResponse = generatedJson.text;
-				emotion = generatedJson.emotion.toLowerCase();
-				state = generatedJson.state;
-				window.state = state;
-				console.log('emotion', emotion);
-				console.log('text', window.aiResponse);
-				console.log('state', generatedJson.state);
-				msgs.push({ role: "assistant", content: window.aiResponse });
-				if (state == 1) {
-					transcript += '\n' + prompts[prompt_index] + JSON.stringify(...msgs) + '\n';
-					console.log(transcript);
-					console.log(contextPrompt + JSON.parse(transcript) + prompts[prompt_index] + emotionPrompt) //to do
-				}
-
-			} catch (e) {
-				console.warn("response is stuctured differently", e);
-			}
-
-		}
-		else {
-			console.log('error', response,);
-		}
-
-	} catch (error) {
-		console.error('Error calling llm', error.message, window.recognizedText);
-	} finally {
-		//window.aiResponse=window.aiResponse.replace(/[?!]/g, '.');
-		generateSpeech(window.aiResponse);
-
-	}
-}
-
-
-function testTTS(text) {
-	generateSpeech(text);
-}
-
-function chat(text) {
-	window.recognizedText = text;
-	getResponse('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');
-}
-
-function resetChat() {
-	window.aiResponse = '';
-	msgs = [
-	];
-	state = 0;
-	window.state = state;
-	window.act_over = false;
-	emotion = 'neutral';
-	window.recognizedText = '';
-	console.log('Chat reset');
-}
-
-function selectPrompt1() {
-	prompt_index = 0;
-}
-function selectPrompt2() {
-	prompt_index = 1;
-}
-function selectPrompt3() {
-	prompt_index = 2;
-}
-window.send_log = send_log;
-window.getResponse = getResponse;
-window.selectPrompt1 = selectPrompt1;
-window.selectPrompt2 = selectPrompt2;
+//window.send_log = send_log;
+//window.getResponse = getResponse;
+//window.selectPrompt1 = selectPrompt1;
+//window.selectPrompt2 = selectPrompt2;
